+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### Next
 - Add long press images to view expanded
-- Add basic forum support
+- Add basic forum support, still in early stages
+- Add quick increment progress to home screen
 - Fix source/licensed by filters not allowing multiple values
 - Fix last added anime to actually show last added